Replace all printing with new OutputFormatter
The OutputFormatter is intended to be a layer at which we
control/dispatch on output types, styles, and apply all of the bells and
whistles which we find necessary.

It replaces all calls to check the output format, and all direct prints.
Accepts several well-known strings to describe textual output formats,
and converters and othersuch for handling inputs which don't match up
with assumptions. Text output can also be completely overwritten with
custom strings, decorated with preambles and epilogs, and generally made
usable without too much complexity.

JSON output is pretty simple, but does accept and use a json_converter
func which transforms non-dict data into a dict. That lets us convert
iterables and othersuch easily.