Add Search index deletion comand
Like the creation command, this uses a (currently) beta API and
therefore is labeled as beta in the helptext.

The intention is for Search to eventually move to a non-`/beta/`
endpoint for this API, at which point the ClI can update to use that
endpoint and drop "BETA" from the helptext.