Implement lazy subcommand loading
All command-to-group attachments are now done lazily, via a dict named
`lazy_subcommands` attached to Group Command objects. The dict records
a name->(config) mapping that allows for lazy loading.

This improves the speed of almost all commands.
All tests pass with no changes.

In the case of a group helptext, the lazy loading will be triggered in
order to fetch command objects (and their short_help attributes), but
it won't cascade down to the second-order children of groups found in
that pass. In the case of completion, only `list_commands` should be
needed, meaning there's no subcommand loading at all because names
are available from the lazy loading mapping.

This also operates in a beneficial way *outside* of the helptext and
completion scenarios: a command like `globus group list` will not need
to parse and load all of the modules for Endpoint ACL management,
Search roles, Timers, or any other non-Groups-related commands.