Bump mindeps for 'cryptography' to 3.3.1
3.3.1 is still an old release of `cryptography`, but the even
older one we were testing has started to be a problem to install/build
on some workstations due to the newer openssl versions shipped with
newer Linux distros.

3.3.1 is actually declared in the setup.py dependencies, so we should
attempt to match that in our testing.

Also, add typing overloads to the coverage exemptions, so that we
don't get false coverage misses on the overload bodies (which cannot
ever execute).