Refine endpointish params decorator usage
Rather than two decorators, go back to a single decorator for both
cases, which splits on a "mode" parameter (Literal[str] type). This
reduces the size and complexity of the code for the decorator by
eliminating the typing overloads while still getting the desired
dispatch behavior.

Additionally, the choice between `--display-name` (option) and
`DISPLAY_NAME` (argument) is made as part of this decoration. This
does slightly complicate usage, as it is now order-dependent with
other positional arguments, but this is nothing new in terms of
argument sharing decorators.

Because the new usage uses Literal, it also pulls in `assert_never`
(which will enter `typing` in py3.11) for exhaustiveness checking,
ensuring that we don't have unhandled Literal values.
This, in turn, means that we need to pull typing-extensions into the
setup.py data for any python < 3.11.

Apply the new decorator to `endpoint create` and `endpoint update`.
This has a small but significant impact on `endpoint create` in that
`--no-default-directory` is no longer a valid option to pass. It would
have served no purpose anyway, since it is mutex with
`--default-directory` and isn't set automatically on create.