Remove mypy from pre-commit
mypy is run under `tox -e mypy` already, in a more reproducible way.
Under `pre-commit`, we run into issues because the same dependency
graph as the real package must be installed, but
`additional_dependencies` does not support this cleanly.

A similar change and range of issues is explained in greater depth in
the globus-sdk, under
  https://github.com/globus/globus-sdk-python/pull/620

Similarly here, the changes allow for local and CI runs on mypy to
continue, but remove it from pre-commit config.