Give Field objects formatting capabilities
Fields contain a `formatter` which can be set to presets via an enum
interface, or can be set to custom objects. To demonstrate this, a
"str" formatter is defined (the default) and a "date" formatter is
defined which replaces the isoformat_to_local conversion function used
by several commands.

Updates to the fields used by timer commands required a tweak to a
test which walked those fields on the assumption that they were all
defined as tuples.