Convert collection commands to use formatters
Wrap the connector_id_to_display_name function in a formatter which
can be used on the relevant field.

Additionally, update the signature for `FieldFormatter.format` to
output `str | None`, so that there is a clear nullary behavior when
formatting fails.

Cleanup and simplify the body of code in globus_cli.utils by moving
functions which were only used by `collection show` into the module
for that command. To support this simplification, the code is changed
to use Field instances only (no tuple-fields). Consequently, the field
filtering can be reduced down to invoking keyfuncs for all fields and
checking the results.
Without any changes to the default keyfunc behavior, this would fail
in some cases. Nested lookups traverse dotted paths and would crash
trying to do a __getitem__ call on a None. Convert the dotted path
traversal code to stop on any None being reached.