Begin converting all fields lists to list[Field]
As a first step in this process, annotate the `formatted_print` fields
argument as `FIELD_LIST_T`, which is a union of several acceptable
types. For two forms which aren't accepted -- `list[Field]` and
`list[tuple[...]]` -- `FIELD_LIST_T` is extended. Beyond that,
anything non-conformant under a mypy run is forced to a list of Field
objects.

This begins the process of eliminating the tuple form from Fields.