Restructure field formatters to parse()->render()
After a discussion about the correct API for FieldFormatter, we have
arrived at a two stage process: parsing and rendering. Parsing is
loading a value and converting it to a desired type, and rendering is
taking the known type and converting it to a string.

This breakdown allows us to express all of the types which we might
want relatively easily. We wrap the `render(parse())` call stack with
a ValueError handler which emits a warning and handles the fallback
case with `str(...)`.

As a related change, remove the FormatName enum. formatter=... is
always passed a formatter instance now.

`glbous task pause-info` was discovered in the course of this work to
have been broken. It is not running during the testsuite, which is how
this wasn't flagged earlier. It is repaired as part of this work and
the extensions to the BoolFieldFormatter which it requires.