Convert Search principal rendering to a formatter
This is a formatter which handles the passing of an AuthClient,
aggregation of data into an IdentityMap, lookups in the map, and
various ways of formatting the different principal types.

It probably doesn't match what other commands are doing, but it could
grow into a replacement for those usages in the future.