Replace `flow list` usage of PrincipalResolver
Instead of PrincipalResolver, with its conditional URN handling, there
are two new formatter classes: `IdentityFormatter` and
`IdentityURNFormatter`.

`flow list` then displays identity IDs using the IdentityURNFormatter
instead of PrincipalResolver.

One thing which needs to change is that the formatter requires an
AuthClient at construction time. In order to support this, the field
definitions for `flow list` are moved into the relevant command.