Use IdentityFormatter for 'collection list'
This required adjustments to the tests. It's not clear how these tests
were passing before, but they were unsafely attempting to pull the
last GCS network request by looking at the last sent request. In fact,
with the identity resolution happening on output, the last call may be
made to Auth. This started failing after the change, so update to have
the tests select more precisely for the request which is relevant.