Attempt to consolidate PrincipalFormatter classes
The end result is not shorter than the prior version in number of
lines, but it aims for a clearer naming scheme and breakdown of
responsibilities such that we only have one notion of a formatter
containing an IdentityMap.

1. A PrincipalFormatter base class defines core operations, based
   around a tuple of strings with the following outline

    parse() -> (principal: str, principal_type: str)
    render(identity_id, "identity") -> username
    render(group_id, "group") -> "Group ({group_id})"
    render(other, _) -> other

2. Various cases are encoded as subclasses of PrincipalFormatter.

    IdentityStrFormatter:
        parse(x) -> (x, "identity")

    IdentityURNFormatter:
        parse("urn:globus:auth:identity:x") -> (x, "identity")
        parse("urn:globus:group:id:x") -> (x, "group")
        parse(x) -> (x, "fallback")

    PrincipalDictFormatter:
        parse({"principal_type": "identity",
               "principal": X}) -> (X, "identity")
        parse({"principal_type": "group",
               "principal": X}) -> (X, "group")
        parse({"principal_type": _,
               "principal": X}) -> (X, "fallback")

3. Special cases are handled by subclassing and overriding

The new structure eliminates parameters to `__init__` for the base
formatter classes. For example, `values_are_urns=False` is not used
because this behavior should be fully covered by the
PrincipalURNFormatter. However, there are still some cases for tuning
formatters for specific commands. In these cases, we can subclass and
override methods which are designed to be overridden, like
`PrincipalFormatter.fallback_rendering`. This keeps the core classes
leaner and simpler while still supporting various special cases.

As a related change here, Globus Search field formatting switches away
from the dict formatter and starts using the PrincipalURNFormatter.
Both the testsuite and live API interactions validate that this
behaves correctly.

4. Better names for formatters
   PrincipalDictFormatter > PrincipalWithTypeKeyFormatter