Add `globus flows delete` to delete flow by ID
In addition to the core change here, a new command, some of the
parsing code is slightly rearranged. `endpoint_id_arg` and
`collection_id_arg` helpers are moved into a dedicated module where
they are joined by `flow_id_arg`. The goal of this adjustment is not
to make major changes, but to better make room for `flow_id_arg`.

The test case relies on globus_sdk._testing data as well as an
explicit mock.