Consolidate "endpointish_setattr_params"
This is a unified helper which should provide parameters for
creating/updating attributes on Endpoint and Collection objects
(the class of entities called "Endpointish" within the globus-cli).

A pre-existing helper `endpointish_create_and_update_params` is
rebranded and expanded for this purposes. Previously, the `create`
vs `update` switch dispatched between different behaviors, and an
option allowed the caller to describe the entity as "collection" or
"endpoint". Now, "collection"|"endpoint" is used for additional
behavioral dispatch, selecting the style of the `--keywords` flag
and the way that verification options are set.