mypy.ini: enumerate modules in 'commands'
Unlike various other manual changes, this was autogenerated by
removing the rule for `commands.*`, running mypy, and putting the
output through a pipeline to generate lines of config.

This ensures that disallow_untyped_defs=false is only set for existing
modules, and those modules are named explicitly in all cases.