Add click annotation checker and begin using it
The annotation checker has specialized knowledge of how click params
map back to types, and customizations to handle the param types defined
in the CLI.

It encodes logic for mapping the parameter list for command objects
back to expected sets of annotated arguments. It then checks
annotations *in a runtime, test context* against those expectations
and raises errors if there are detected deviations.

The new helper is applied to several recently annotated commands and
has the following basic results:
- `--notify` needs to be moved to a custom param type so that it can be
  understood by the annotation checker
- `--batch` was annotated incorrectly in `globus transfer`; missing
  `| None` in the annotation

Importantly, `--notify` heralds a potential sweeping change, that any
custom callback may require special handling. As an alternative, we
could introduce custom subclasses of Option and Argument which allow
us to specify known types as part of the runtime specification.

The tests only run on py3.10+ in order to allow use of
`types.UnionType` and have other useful features.