Add tests for `--notify` and fix a bug
The latest change which converted the `--notify` callback to a type
introduced a regression. Missing `--notify` would be passed as `None`,
rather than the expected `{}`. To fix, introduce a callback which
simply converts `None` to `{}` explicitly. The logic for how
"non-nullable options" in the annotation checker needs updating to
handle this corectly, and the new tests are using the callback as a
simple way to exercise its behavior.