Refine click param type checker and add self-tests
This changeset refines the click parameter type checker to be more
readable and simpler. Primarily, it factors out the logic of
converting a click.Parameter object into a type. Additionally, it
reduces the walks of parameters down to a single loop for improved
readability.

As part of the refactor, unit tests for the param checker itself are
introduced. These help guarantee that the checker runs correctly on
various "simple" cases. In the course of this testing, some minor bugs
with the checker were revealed:
- `multiple` params should always be `tuple[X, ...]`, they do not
  default to `None`
- `multiple` params can take the form of an argument with `nargs=-1`,
  account for this

This applies a fix which was noted in review, to union-ize the
multiple types seen under a multiple option and include them in the
tuple type.

An additional fix to the `globus transfer` command is here identified:
the type for `exclude` does not include `None`. Setting a default of
`None` has no effect. The default is `()`. The annotation is
corrected, and a relevant line of source is also fixed.