Refactor NotificationParamType for clarity
The main body of `NotificationParamType`, the `convert()` method, was
hard to read because it created too many variables in a single scope
and rebound the type of one of those variables.

This rewrite (proposed by @kurtmckee) makes the main body clearer by
avoiding changing the type of `value`, and dropping the `on` and `off`
bools. It also improves the behavior when an invalid option is given
by calling it out by name in the error message.

Additionally, fix type annotations on the notify param type and on the
none-to-bool callback.

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>