Simplify `--notify` shell complete; fewer paths
There are fewer conditional and branching points in this rewrite of
the functionaltiy, inspired heavily by comments from @kurtmckee .

A minor simplification: collect results from `split(",")` using
left-hand-side `*args` capturing.

More significantly: condense all of the logic after the `"," in str`
check down to almost a single path. The logic is now trying to build a
set of possible completions for the last post-comma fragment,
*including* the case when that fragment is `""`. There are a couple of
corner cases which need careful handling, but overall the logic is
simpler and less nested.