Use entity_types in endpointish_setattr_params
Several refactors are applied here, primarily to the interface of
endpointish_setattr_params. entity_types (a tuple of applicable entity
types) is a required argument. It is both descriptive, in that it
tells us how a command is designed to be used, and prescriptive in
that it sets various behaviors for parsing.

entity_types replaces dispatching between `collection` and `endpoint`
as a way of setting options. These are retained as a purely
descriptive field which appears in helptext.

Internally, the endpointish_setattr_params decorator is also improved.
It now converts more clearly from entity_types to various flag
settings, and then produces a singular, well-formed, decorator. It
also gains the capability to set overrides in a dict, to control these
settings where the deduction could be mistaken.

Apply the refined endpointish_setattr_params decorator to
'endpoint update', carefully retaining all existing behaviors, and
rewriting away from `**kwargs` capturing usage. The click type checker
can now run on the result successfully.