Add testing for 'endpoint is-activated'
This command was not covered by the testsuite prior to this addition.

Add a new set of functional tests and remove the disallow_untyped_defs
setting for this command. List it in the click annotation checks.

Several small tweaks are made to 'is_activated.py' -- modern use of
f-strings, fix missing annotations, etc.
The output is changed in very minor ways, for text output only.