endpointish_setattr_params: remove 'entity_types'
Replace 'entity_types' and deduction with a combination of default
behaviors and allowing for overrides.

That is, instead of asking for "options appropriate to GCP Mapped
Collections", usage instead says (for 'gcp create mapped') "standard
create options, with keywords as a string".

The transformation results in a smaller internal surface area to
maintain and, as it turns out, simpler usage.

Several annotations are corrected here as well.