Introduce AnnotatedOption to check verify types
The verify choice option used in `endpointish_setattr_params` does not
produce a literal string choice, but passes through a callback which
is guaranteed to return a `dict[str, bool]`.

The annotation checker could not understand this, and therefore could
not be applied to the methods which use `verify_style="choice"`.
To resolve, introduce a custom click.Option class, AnnotatedOption,
which allows annotations to be attached to the Option and retrieved
for the check.

A significant wrinkle emerges, which is that `dict != typing.Dict`,
and therefore we need to be careful about compatibility. To handle
this case, the dict type for the verify choice option is selected
*at runtime* between these two values based on a dispatch over python
versions.

With these changes in place and the type of `verify` now checkable,
apply the type checker to the `globus gcp create` commands.