Distribute param type annotations across params
More deeply integrate our type annotation support into click
parameter types by introducing `AnnotatedParamType`, a subclass of
`ParamType` which adds a type annotation callback method.

Similar to AnnotatedOption, this allows us to distribute the burden of
documenting the types associated with click classes across our
codebase. AnnotatedParamType objects have a method which deduces types
from the param-type instance. For example, it can check on our
timedelta type if we are set to use seconds (`int`) or a delta
(`datetime.timedelta`).

With this change in place, a refactor of the click param type helper
is possible, eliminating case-wise logic in favor of dispatch back
down to the various `AnnotatedParamType`s.