Run the annotation checker on 'endpoint create'
Add the 'endpoint create' command to the click annotation checker and
then make the necessary adjustments.

The `one_use_option` decorator intercepts arguments and swaps out
types in a way which confuses the checker when it is used to declare a
flag. In order to support this, when no explicit annotation is
included and `one_use_option` is declaring a flag, it will set the
annotation to `bool`.

In a one case, an annotation is being set to `tuple[UUID, str]` and
at runtime. `mypy` flags this usage incorrectly, claiming that
`tuple` can't support the given type args. To work around this,
simply apply the appropriate type-ignore comment for now.

Significantly, expand the `**kwargs` list in the endpoint create
command, and ensure that all of the arguments are passed onwards
correctly to helpers, etc.