Namespace and collect endpointish_params better
The parameter grouping is now done under a class which acts as a
namespace, providing the decorators as classmethods.
`@endpointish_params.create` and `@endpointish_params.update` are the
two primary decorators. The collected options from the
`globus endpoint create|update` commands are now included as well,
under a name of `legacy_transfer_params`.