one_use_option flags must set type to bool
Prior to this change, `one_use_option` would detect a flag being
converted to a count option and set the type to `bool`. Now, the type
is not set automatically ever (fewer different modes for us to
maintain), but these usages need to pass `type_annotation=bool`. There
are only three such usages in two commands.