Move user_message{,_link} into shared options
These parameters are now part of the shared create/update parameters
for endpointish objects. They can be set on collections and host
endpoints, but are noted in documentation as not being valid on
shares. As such, they are not included in the options for
`globus gcp create guest`.

In order to achieve the above semantics, the shared option decorators
have gained a new ability, to accept a tuple of strings to denote
options which should be omitted or "skipped".
The guest collection creation command passes

    skip=("user_message", "user_message_link")

and this is enough to signal that these two options are skipped. To do
this, a mild refactor is applied, turning a list of decorators into a
map of names to decorators. The final step, when they are applied, is
a filter and conversion to a list.

closes #476