Update to globus-sdk v3.16.0
And handle the change to session_required_policies between 3.15.1 and
3.16.0 .

Because this is now passed around as a list, several changes are made
here to harmonize the way that this list of IDs is handled. The
exception hook now expects the SDK to pass the list[str], but the
`sesion update` command has also been changed to explicitly parse the
input as a CommaDelimitedList and then comma-join it for use in Auth.

The existing tests, minimally updated, exercise this behavior.

Also ensure that the policy IDs are quoted when printed in a command,
as a best-practice for how commands should be generated for users.
Single-quotes ensure that we get verbatim values in all cases, even if
they are not strictly required in this one.