Improve the formatting of failed test output
The `run_line` fixture tries to present all of the behavior of a
command which fails in the form of an exception message.
But the message formatting is currently hard to read, and it can
easily be improved with some better whitespace. This doesn't impact
any tests passing/failing, as the output only appears in the case
of a failing test.