Add "well_known_config" internal APIs for storage
Define clear helpers for interacting with the config storage. This
will enable use from outside of the narrow space of login/logout
functionality, but also defines a better interface for other future
uses.

We have already encountered some need for this kind of change in order
to support the Timer commands.

Because the API defined here relies on `Literal` checking for safety,
we need to ensure that all of its usage sites are properly annotated.
Therefore, annotate several functions which previously were not
annotated.