Add dynamic dependencies to Timer scope
Timer requires that we build a dynamic dependency for the
TransferAP->Transfer dependency chain. Therefore, we need sufficient
tooling to build this dependency chain *and* to build the dependency
chain with data_access scopes at the leaves.

To support this, a `scopes` module is added to the `login_manager`
subpackage and it contains a helper for building Timer scopes as well
as a couple of useful constants.

This also introduces the notion of "contract versions" to the login
manager, but without useful enforcement logic at present. The idea of
the contract version is that it tracks when our expectations/needs for
a scope change in a way which we don't believe can be detected without
an explicit flag.