Add scope "contract version" enforcement
Even with our recent refinements to the login manager to capture scope
requirements, there are still scope configuration changes which cannot
be captured by our detection logic.

In order for the CLI to actively attempt to detect these conditions,
we have two options:
- make more aggressive use of the Consents API
- add some internal tracking so that we know when our scopes change

The existing checks we've added for scope strings are still good and
useful for the majority of cases. However, for dynamic scope
dependencies, we need one of these mechanisms. Relevantly, the Timer
scope needs to be used with dynamic dependencies attached to it, and
this cannot be detected by the mechanisms we have today.

Using the Consents API would mean that when checking if the user is
logged in, we need to make a call out to Auth. This extra roundtrip is
part of a growing speed penalty on "launch" of a command that we would
ideally avoid. Using local storage is marginally more complex but much
faster because we just need to make a roundtrip to disk to get info on
what we last requested.

Rather than storing the dynamic dependencies or other specific data
which might not handle other scenarios in the future, this introduces
a generic "scope contract version" number, which is declared starting
at 1 (with existing tokens implicitly under contract version 0). We
store this in the config table of tokenstorage and mark all tokens
with our current version number in a `resource_server: integer`
mapping. The login checks can then simply compare the known contract
version against data from disk and thereby get a comparison between
our current scope requirements and the requirements at the time that
the user did a login flow for that token.

In order to implement this, the tokenstorage wrapping module needs to
define a custom SQLiteAdapter subclass. And to maintain our deferred
import performance tuning in the CLI, the subclass needs to be defined
dynamically inside of a helper. Once that is taken care of, however,
the implementation is "relatively straightforward".

The CURRENT_SCOPE_CONTRACT_VERSION is a constant initialized with a
value of 1, and is what will be written to disk.