Rewrite the form of token/scope requirements
Define a new class, `_CLIScopeRequirements` (internal) and a singleton
instance of it, `CLI_SCOPE_REQUIREMENTS` which extract the notions
previously encoded by `LoginManager.STATIC_SCOPES` and several other
attributes of the LoginManager class.

Because this changes how we address resource servers by known "nice"
names, it touches almost every command in the CLI with an update to
their login requirements. Most of these changes are purely cosmetic.

The LoginManager now depends upon the CLI_SCOPE_REQUIREMENTS and
interacts with it, so the problems of one and the other are somewhat
separated.

CLI_SCOPE_REQUIREMENTS can address information by "nice names" like
"transfer" and by full resource server names. Each service is listed
with their min_contract_version for the scope contract versions idea.
The data within the mapping is structured using a TypedDict, which
gives us some level of type-safety in our accesses, and the patching
for login manager tests can simply rewrite the full suite of data
using ordinary dicts to get desired test-time semantics out of the
system.