Add version smudging script for dev releases
This is a script which adds a `devX` version number to the end of the
current version string. It takes a `--revert` flag which will return
the version number to a 'X.Y.Z' version assuming that the smudger was
used to modify the version in the first place.

This tool is meant to support automated testpypi releases using
modified version numbers.